requirements.txt
fastapi==0.115.*
uvicorn==0.30.*
pydantic==2.*
SQLAlchemy==2.*
python-dotenv==1.*
langchain==0.3.*
langchain-openai==0.2.*
streamlit==1.36.*
requests==2.32.*
pytest==8.0.*
pytest-asyncio==0.23.*
httpx==0.26.*
pytest-mock==3.12.*
faker==22.0.*
